:recycle: Renames admin user function
