+++ packages/backend/aws/src/cognito/index.ts
@@ -1,7 +1,6 @@
 export * from "./cognito-configure";
-export * from "./create-cognito-user";
+export * from "./create-cognito-admin-user";
 export * from "./delete-cognito-user";
 export * from "./update-cognito-user-groups";
-export * from "./create-cognito-admin-user";
 export * from "./create-cognito-basic-user";
 export * from "./verify-basic-user-account";
